#65 [refactor] : Enum의 패키지를 이동한다
