+++ src/main/java/side/onetime/domain/enums/Category.java
@@ -1,0 +1,5 @@
+package side.onetime.domain.enums;
+
+public enum Category {
+    DATE, DAY
+}
