--- src/main/java/side/onetime/global/common/constant/Category.java
@@ -1,5 +1,0 @@
-package side.onetime.global.common.constant;
-
-public enum Category {
-    DATE, DAY
-}
